# backend/public/api-docs/openapi.yaml
openapi: 3.0.0
info:
  title: Sassy CRM API
  description: Modern headless CRM API for software sales teams with AI-powered features
  version: 1.0.0
  contact:
    name: API Support
    email: support@sassycrm.com
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.sassycrm.com/api
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /crm/dashboard/pipeline:
    get:
      tags: [Dashboard]
      summary: Get pipeline data
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        200:
          description: Pipeline data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PipelineData'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Opportunity:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        opportunity_type:
          type: string
        account_id:
          type: string
        campaign_id:
          type: string
        lead_source:
          type: string
        amount:
          type: number
          format: float
        amount_usdollar:
          type: number
          format: float
        currency_id:
          type: string
        date_closed:
          type: string
          format: date
        next_step:
          type: string
        sales_stage:
          type: string
        probability:
          type: integer
        description:
          type: string
        assigned_user_id:
          type: string
        date_entered:
          type: string
          format: date-time
        date_modified:
          type: string
          format: date-time
...
    PipelineData:
      type: object
      additionalProperties:
        type: object
        properties:
          count:
            type: integer
          value:
            type: number
          opportunities:
            type: array
            items:
              $ref: '#/components/schemas/Opportunity'
...